requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
PyYAML>=6.0
PyPDF2>=3.0.0
openai>=1.0.0
anthropic>=0.7.0
ollama>=0.0.1
